requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.0.3
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.5
PyJWT==1.6.4
python-coveralls==2.9.1
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
